rename model file to PreenlistmentCourse.js
